code modified for the test branch; removed requirement for publicized assemblies and instead use BepInEx.AssemblyPublicizer to publicize them during referencing directly from the Valheim install folder

+++ ChebsValheimLibrary/ChebsValheimLibrary/Minions/AI/WoodcutterAI.cs
@@ -241,7 +241,7 @@ namespace ChebsValheimLibrary.Minions.AI
 
             var hitData = new HitData();
             hitData.m_damage.m_chop = axeItem.m_itemData.m_shared.m_damages.m_chop;
-            hitData.m_toolTier = axeItem.m_itemData.m_shared.m_toolTier;
+            hitData.m_toolTier = (short)axeItem.m_itemData.m_shared.m_toolTier;
             destructible?.Damage(hitData);
 
             _chopping = false;
